+++ build.sbt
@@ -1,4 +1,4 @@
-name          := "Rush_Hour"
+name          := "Dominion"
 organization  := "de.htwg.se"
 version       := "0.0.1"
 scalaVersion  := "2.12.4"
