+++ src/main/scala/de/htwg/se/Dominion/Dominion.scala
@@ -1,8 +1,8 @@
-package de.htwg.se.Rush_Hour
+package de.htwg.se.Dominion
 
-import de.htwg.se.Rush_Hour.model.Player
+import de.htwg.se.Dominion.model.Player
 
-object Rush_Hour {
+object Dominion {
   def main(args: Array[String]): Unit = {
     val student = Player("Jakob Strakhof")
     println("Hello, " + student.name)
