+++ src/main/scala/de/htwg/se/Dominion/model/ModelWorksheet.sc
@@ -1,0 +1,3 @@
+object printProject {
+  println("Project: Dominion")
+}
