+++ src/test/scala/de/htwg/se/Dominion/model/PlayerSpec.scala
@@ -1,4 +1,4 @@
-package de.htwg.se.Rush_Hour.model
+package de.htwg.se.Dominion.model
 
 import org.scalatest._
 import org.junit.runner.RunWith
@@ -7,12 +7,12 @@ import org.scalatest.junit.JUnitRunner
 @RunWith(classOf[JUnitRunner])
 class PlayerSpec extends WordSpec with Matchers {
   "A Player" when { "new" should {
-    val player = Player("Your Name")
+    val player = Player("Jakob Strakhof")
     "have a name"  in {
-      player.name should be("Your Name")
+      player.name should be("Jakob Strakhof")
     }
     "have a nice String representation" in {
-      player.toString should be("Your Name")
+      player.toString should be("Jakob Strakhof")
     }
   }}
